house cleaning for 32-bit compilation
Avoid using size_t as an implied 64-bit value, as on 32-bit systems
that is not correct.

Reorganized the headers so that we can reuse bitops without gmp.

Made the add_cxx_compiler_options portable.
