+++ cmake/P4CUtils.cmake
@@ -7,11 +7,11 @@ MACRO (add_cxx_compiler_option option)
   string (REPLACE "-" "" escaped_option ${escaped_option1})
   set (_success "_HAVE_OPTION_${escaped_option}_")
   set (__required_flags_backup "${CMAKE_REQUIRED_FLAGS}")
-  set (CMAKE_REQUIRED_FLAGS "${CMAKE_REQUIRED_FLAGS} ${P4C_CXX_FLAGS}")
+  set (CMAKE_REQUIRED_FLAGS "${CMAKE_REQUIRED_FLAGS} ${${PROJECT_NAME}_CXX_FLAGS}")
   check_cxx_compiler_flag (${option} ${_success})
   set (CMAKE_REQUIRED_FLAGS ${__required_flags_backup})
   if (${_success})
-    set (P4C_CXX_FLAGS "${P4C_CXX_FLAGS} ${option}")
+    set (${PROJECT_NAME}_CXX_FLAGS "${${PROJECT_NAME}_CXX_FLAGS} ${option}")
   endif (${_success})
 endmacro (add_cxx_compiler_option)
