+++ lib/bitvec.h
@@ -1,5 +1,5 @@
 /*
-Copyright 2013-present Barefoot Networks, Inc. 
+Copyright 2013-present Barefoot Networks, Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -45,13 +45,10 @@ limitations under the License.
 #if defined(__GNUC__) || defined(__clang__)
 /* use builtin count leading/trailing bits of type-approprite size */
 static inline int builtin_ctz(unsigned x) { return __builtin_ctz(x); }
-static inline int builtin_ctz(unsigned long x) { return __builtin_ctzl(x); }
 static inline int builtin_ctz(unsigned long long x) { return __builtin_ctzll(x); }
 static inline int builtin_clz(unsigned x) { return __builtin_clz(x); }
-static inline int builtin_clz(unsigned long x) { return __builtin_clzl(x); }
 static inline int builtin_clz(unsigned long long x) { return __builtin_clzll(x); }
 static inline int builtin_popcount(unsigned x) { return __builtin_popcount(x); }
-static inline int builtin_popcount(unsigned long x) { return __builtin_popcountl(x); }
 static inline int builtin_popcount(unsigned long long x) { return __builtin_popcountll(x); }
 #endif
 
@@ -478,6 +475,5 @@ inline bitvec operator^(bitvec &&a, const bitvec &b) {
     bitvec rv(std::move(a)); rv ^= b; return rv; }
 inline bitvec operator-(bitvec &&a, const bitvec &b) {
     bitvec rv(std::move(a)); rv -= b; return rv; }
-
 
 #endif  // P4C_LIB_BITVEC_H_
