+++ lib/stringify.h
@@ -20,7 +20,7 @@ limitations under the License.
 #define P4C_LIB_STRINGIFY_H_
 
 #include <stdint.h>
-#include "gmputil.h"  // for mpz_class
+#include <gmpxx.h>  // for mpz_class
 #include "cstring.h"
 #include "stringref.h"
